Extract browser-only functions into Utils
This will make it easier to mock out these methods in tests to use Node apporaches to these problems
